+++ src/platform_impl/windows/window_state.rs
@@ -85,6 +85,13 @@ bitflags! {
 
         const MINIMIZED = 1 << 12;
 
+        const IGNORE_MOUSE_EVENT = 1 << 14;
+
+        const FULLSCREEN_AND_MASK = !(
+            WindowFlags::DECORATIONS.bits |
+            WindowFlags::RESIZABLE.bits |
+            WindowFlags::MAXIMIZED.bits
+        );
         const EXCLUSIVE_FULLSCREEN_OR_MASK = WindowFlags::ALWAYS_ON_TOP.bits;
         const NO_DECORATIONS_AND_MASK = !WindowFlags::RESIZABLE.bits;
         const INVISIBLE_AND_MASK = !WindowFlags::MAXIMIZED.bits;
@@ -202,7 +209,11 @@ impl WindowFlags {
             style |= WS_VISIBLE;
         }
         if self.contains(WindowFlags::ON_TASKBAR) {
+            // show taskbar icon
             style_ex |= WS_EX_APPWINDOW;
+        } else {
+            // not show taskbar icon, https://docs.microsoft.com/en-us/windows/win32/winmsg/extended-window-styles
+            style_ex |= WS_EX_TOOLWINDOW;
         }
         if self.contains(WindowFlags::ALWAYS_ON_TOP) {
             style_ex |= WS_EX_TOPMOST;
@@ -219,6 +230,9 @@ impl WindowFlags {
         if self.contains(WindowFlags::MAXIMIZED) {
             style |= WS_MAXIMIZE;
         }
+        if self.contains(WindowFlags::IGNORE_MOUSE_EVENT) {
+            style_ex |= WS_EX_TRANSPARENT | WS_EX_LAYERED;
+        }
 
         style |= WS_CLIPSIBLINGS | WS_CLIPCHILDREN | WS_SYSMENU;
         style_ex |= WS_EX_ACCEPTFILES;
@@ -228,7 +242,6 @@ impl WindowFlags {
         ) {
             style &= !WS_OVERLAPPEDWINDOW;
         }
-
         (style, style_ex)
     }